Separate embedded mqtt broker status from client_config
To have the embedded mqtt broker work with a custom config, we need to
separate the client config bit from the broker status bits.

In lower layers they where already separate but in higher layers they
where merged, but in a faulty way. That caused HA to think the broker
setup failed, when it didn't generate a client_config.

When running a embedded mqtt broker with some custom config, one needs
also to provide a client config which works with that broker setup.